+++ src/core.ts
@@ -163,7 +163,7 @@ export abstract class APIClient {
     maxRetries = 2,
     timeout = 60000, // 1 minute
     httpAgent,
-    fetch: overridenFetch,
+    fetch: overriddenFetch,
   }: {
     baseURL: string;
     maxRetries?: number | undefined;
@@ -176,7 +176,7 @@ export abstract class APIClient {
     this.timeout = validatePositiveInteger('timeout', timeout);
     this.httpAgent = httpAgent;
 
-    this.fetch = overridenFetch ?? fetch;
+    this.fetch = overriddenFetch ?? fetch;
   }
 
   protected authHeaders(opts: FinalRequestOptions): Headers {
@@ -268,6 +268,7 @@ export abstract class APIClient {
     options: FinalRequestOptions<Req>,
     { retryCount = 0 }: { retryCount?: number } = {},
   ): { req: RequestInit; url: string; timeout: number } {
+    options = { ...options };
     const { method, path, query, headers: headers = {} } = options;
 
     const body =
@@ -278,9 +279,9 @@ export abstract class APIClient {
 
     const url = this.buildURL(path!, query);
     if ('timeout' in options) validatePositiveInteger('timeout', options.timeout);
-    const timeout = options.timeout ?? this.timeout;
+    options.timeout = options.timeout ?? this.timeout;
     const httpAgent = options.httpAgent ?? this.httpAgent ?? getDefaultAgent(url);
-    const minAgentTimeout = timeout + 1000;
+    const minAgentTimeout = options.timeout + 1000;
     if (
       typeof (httpAgent as any)?.options?.timeout === 'number' &&
       minAgentTimeout > ((httpAgent as any).options.timeout ?? 0)
@@ -309,7 +310,7 @@ export abstract class APIClient {
       signal: options.signal ?? null,
     };
 
-    return { req, url, timeout };
+    return { req, url, timeout: options.timeout };
   }
 
   private buildHeaders({
@@ -337,14 +338,21 @@ export abstract class APIClient {
       delete reqHeaders['content-type'];
     }
 
-    // Don't set the retry count header if it was already set or removed through default headers or by the
-    // caller. We check `defaultHeaders` and `headers`, which can contain nulls, instead of `reqHeaders` to
-    // account for the removal case.
+    // Don't set theses headers if they were already set or removed through default headers or by the caller.
+    // We check `defaultHeaders` and `headers`, which can contain nulls, instead of `reqHeaders` to account
+    // for the removal case.
     if (
       getHeader(defaultHeaders, 'x-stainless-retry-count') === undefined &&
       getHeader(headers, 'x-stainless-retry-count') === undefined
     ) {
       reqHeaders['x-stainless-retry-count'] = String(retryCount);
+    }
+    if (
+      getHeader(defaultHeaders, 'x-stainless-timeout') === undefined &&
+      getHeader(headers, 'x-stainless-timeout') === undefined &&
+      options.timeout
+    ) {
+      reqHeaders['x-stainless-timeout'] = String(options.timeout);
     }
 
     this.validateHeaders(reqHeaders, headers);
@@ -508,18 +516,22 @@ export abstract class APIClient {
 
     const timeout = setTimeout(() => controller.abort(), ms);
 
-    return (
-      this.getRequestClient()
-        // use undefined this binding; fetch errors if bound to something else in browser/cloudflare
-        .fetch.call(undefined, url, { signal: controller.signal as any, ...options })
-        .finally(() => {
-          clearTimeout(timeout);
-        })
-    );
-  }
+    const fetchOptions = {
+      signal: controller.signal as any,
+      ...options,
+    };
+    if (fetchOptions.method) {
+      // Custom methods like 'patch' need to be uppercased
+      // See https://github.com/nodejs/undici/issues/2294
+      fetchOptions.method = fetchOptions.method.toUpperCase();
+    }
 
-  protected getRequestClient(): RequestClient {
-    return { fetch: this.fetch };
+    return (
+      // use undefined this binding; fetch errors if bound to something else in browser/cloudflare
+      this.fetch.call(undefined, url, fetchOptions).finally(() => {
+        clearTimeout(timeout);
+      })
+    );
   }
 
   private shouldRetry(response: Response): boolean {
@@ -957,8 +969,8 @@ export const safeJSON = (text: string) => {
   }
 };
 
-// https://stackoverflow.com/a/19709846
-const startsWithSchemeRegexp = new RegExp('^(?:[a-z]+:)?//', 'i');
+// https://url.spec.whatwg.org/#url-scheme-string
+const startsWithSchemeRegexp = /^[a-z][a-z0-9+.-]*:/i;
 const isAbsoluteURL = (url: string): boolean => {
   return startsWithSchemeRegexp.test(url);
 };