+++ src/resources/video/delivery-usage.ts
@@ -38,7 +38,7 @@ export interface DeliveryReport {
   asset_duration: number;
 
   /**
-   * @deprecated: This field is deprecated. Please use `asset_video_quality` instead.
+   * @deprecated This field is deprecated. Please use `asset_video_quality` instead.
    * The encoding tier that the asset was ingested at.
    * [See the video quality guide for more details.](https://docs.mux.com/guides/use-video-quality-levels)
    */