+++ tests/api-resources/data/real-time.test.ts
@@ -73,7 +73,7 @@ describe('resource realTime', () => {
         'current-concurrent-viewers',
         {
           dimension: 'asn',
-          filters: ['string', 'string', 'string'],
+          filters: ['string'],
           order_by: 'negative_impact',
           order_direction: 'asc',
           timestamp: 0,
@@ -108,7 +108,7 @@ describe('resource realTime', () => {
     await expect(
       client.data.realTime.retrieveHistogramTimeseries(
         'video-startup-time',
-        { filters: ['string', 'string', 'string'] },
+        { filters: ['string'] },
         { path: '/_stainless_unknown_path' },
       ),
     ).rejects.toThrow(Mux.NotFoundError);
@@ -139,7 +139,7 @@ describe('resource realTime', () => {
     await expect(
       client.data.realTime.retrieveTimeseries(
         'current-concurrent-viewers',
-        { filters: ['string', 'string', 'string'], timestamp: 0 },
+        { filters: ['string'], timestamp: 0 },
         { path: '/_stainless_unknown_path' },
       ),
     ).rejects.toThrow(Mux.NotFoundError);